Extract LCM-bypass fallback handling into a mixin (WS5 seam)
Move the 14 methods that handle sessions bypassing LCM context
management -- bypass detection, host-fallback-compressor mirroring,
and the deterministic bypass tail-compaction -- verbatim out of
engine.py into a new BypassMixin in bypass.py. LCMEngine mixes it in
before ContextEngine, so `self` still resolves every reference; call
sites, shared state, and method names are unchanged.

engine.py -474 lines. The two imports the block solely used
(`importlib`, `_assistant_tool_call_ids`) are dropped from engine.py
and declared in bypass.py.

Behaviour-neutral: the moved method bodies are byte-identical to
main (AST-slice compare, 0 diffs). ruff clean; full suite
1575 passed / 12 xfailed (unchanged from the v0.19.0 baseline).
